Remove refresh_mode option from mongoid.yml
As far as I can tell this option has no effect. It's not an option in
Mongoid and it doesn't appear to be an option in the underlying mongo
gem either.

To test, I added this option to the development group in mongoid.yml and
started the Rails console. I see the following warning:

> refresh_mode is not a valid option for Mongo::MongoClient

I have also searched the alphagov repos on GitHub for 'refresh_mode'[1]
and the only result is for this manuals-publisher repo.

The setting was added in c371b95c14bb1ada89dc6fd31afaefdf3a154f42 but
the commit note doesn't explain anything about it. In the comments on
that commit[2], the original author has explained that the setting was
copied from alphagov-deployment and that there's no explanation in that
repo either.

Removing this should make it easier to upgrade Mongoid in a separate
branch.

[1]: https://github.com/search?q=org%3Aalphagov+refresh_mode&type=Code&utf8=%E2%9C%93
[2]: https://github.com/alphagov/manuals-publisher/commit/c371b95c14bb1ada89dc6fd31afaefdf3a154f42