Update indexes in SectionEdition for Mongoid 3
This avoids the following problem when running `rake`:

    Problem:
      Invalid index specification on SectionEdition: section_id, {}
    Summary:
      Indexes in Mongoid are defined as a hash of field name and
      direction/2d pairs, with a hash for any additional options.
    Resolution:
      Ensure that the index conforms to the correct syntax and has the
      correct options.

The `index` method now takes a field name and direction where it
previously took a name only.

Although the `index` macro in Mongoid 2.8.1 only takes a `name`
parameter, that was being converted to a name/direction pair in
`Collection#parse_index_spec`[1] of the underlying mongo-ruby-driver
Gem. The direction was being set to 1 (ascending) so I'm confident that
we're retaining the same behaviour in this commit.

This change to the `index` macros is mentioned in the Changelog for
3.0.0[2]

> Indexing syntax has changed. The first parameter is now a hash of
> name/direction pairs with an optional second hash parameter for
> additional options.

[1]: https://github.com/mongodb/mongo-ruby-driver/blob/1.x-stable/lib/mongo/collection.rb#L1121
[2]: https://github.com/mongodb/mongoid/blob/master/CHANGELOG.md#300