Inline creation of SectionRepository in SectionAssociationMarshaller
The SectionAssociationMarshaller is always created with the same
`section_repository_factory` argument, and this factory always returns
the same SectionRepository (instantiated with the manual used in the
SectionAssociationMarshaller). I think it is clearer to just
instantiate SectionRepository inline in this class.

This also removes the last reference to
`RepositoryRegistry#section_repository_factory` so that method is also
removed.