Extract RepositoryRegistry#scoped_manual_repository into a class
The RepositoryRegistry#scoped_manual_repository builds an instance of
ManualRepository with a given marshallers and collection argument. The
marshallers argument never changes, but the collection argument
depends on where the method is called from.

To help refactor this, I'm extracting the creation of this into its
own class, which I've called ScopedManualRepository to match the
original method name (although I think we may find a better name
later). This will allow me to remove the indirection in the creation
of these through the `RepositoryRegistry` in a future commit.