Move DocumentHelpers#check_for_javascript_usage_error -> ManualHelpers
This method is used in the context of both manuals and sections.
However, it's the last method defined in `DocumentHelpers` and there
are already methods in `ManualHelpers` which relate to both manuals
and sections, so it doesn't seem worth keeping `DocumentHelpers`
around.

My preference is to extract the relevant methods out of `ManualHelpers`
into a new `SectionHelpers` module.