Inline the decorator argument in SectionAssociationMarshaller
This class is only ever instantiated with the same decorator argument
so it can be inlined into the class.

This has made the tests a little more verbose as we can no longer
simply test that the decorator proc has been called. On the other
hand, we can now be sure that the correct decorators have been
called. It's not clear to me if these decorators really need to be
this complicated, perhaps they can be simplified and incorporated into
the SectionAssociationMarshaller (or another) class at a later date.