Use MONGODB_URI env var in Mongoid config
Now that the application is using Ruby v2.2.4, it seems as if Mongoid
can parse URIs containing multiple comma-separated hosts.

We're hoping that this will allow us to re-apply the changes in #928 to
upgrade Mongoid from v2.5.2 to v2.8.1 which we had to revert in #935,
because of a problem in the integration deployment - specifically
during the pre-compilation of assets which happens to be the first time
the production Mongoid configuration is used.

Fixes #930.