Inline section_builder arg into ManualWithSections constructor
`VersionedManualRepository` used to raise an exception if you tried to
add a new section in order to enforce the fact that its manual is
intended to be read-only. I've made a pragmatic decision to drop this
constraint, because I don't think it's hugely important and the
behaviour was not being tested.