Delegate from Manual to ManualWithSections
Eventually I'm hoping to collapse `ManualWithSections` into `Manual`.
As a first step, this commit constructs an instance of
`ManualWithSections` for every instance of `Manual` and delegates the
relevant methods to the `ManualWithSections` instance. All the other
places in the code which were instantiating a `ManualWithSections` now
just need to instantiate a `Manual`. For the moment these places also
need to explicitly set the `sections` and `removed_sections` attributes
of the `Manual`.

Previously `ManualWithSections` intercepted calls to `#publish` before
they reached `Manual`, but not that the delegation order is reversed,
I've had to rename `Manual#publish` to `#__publish__` so that calls to
`#publish` are still intercepted by `ManualWithSections` and the latter
can still call `Manual#__publish__`. I hope to simplify this in a later
commit.

It might seem a bit confusing that the `ManualWithSections` is still
delegating to `Manual` as well as being delegated *to* by the `Manual`,
but I hope to remove the former is a later commit.