Use Manual.all vs ScopedManualRepository.all in ManualHelpers
Note that I've changed the implementation of `Manual.all` to use
`ScopedManualRepository` vs `ManualRepository`, because the
`ManualHelper` method needs manuals to be loaded with their sections
which is what the former gives us via the
`SectionAssociationMarshaller`.

Note that I was initially a bit hesitant to do this, because I wasn't
sure what the `ManualPublishTaskAssociationMarshaller` did, but I now
understand that it just loads a `ManualPublishTask` for the relevant
manual - any change of state is done by calling methods on the task
itself. So I'm happy that this change is safe. It also has the benefit
of bringing `Manual.all` into line with other `Manual` methods.

The only downside I can see is that manuals will be loaded with their
sections everywhere including on the manuals index page, which might
mean a deterioration in performance. However, I think that's better
address separately if it happens.