Extract Manual.current_versions
This follows on from #953 & #954 and is a further effort at reducing
the number of places which depend on the various manual repositories.
In this case, I'm specifically dealing with `VersionedManualRepository`.
There are still some occurrences left to deal with,
e.g. `ManualRepository::NotFoundError`.

* Use `Manual.find` followed by `Manual#current_versions` in place of
`VersionedManualRepository.get_manual` in `RepublishManualService` and
`ManualRelocator`.

* Supply a `context` to `RepublishManualService` constructor to make it
more consistent with other services.

* Use `User.gds_editor` to supply the relevant context to
`ManualRelocator` and `ManualsRepublisher`.

* Remove the now redundant `VersionedManualRepository.get_manual` and
`RepublishManualService#manual_repository` methods.