Inline ManualPublishTaskAssociationMarshaller#load
This two-line method can be safely inlined in the ManualRepository as
it is not used elsewhere in the codebase.

I've changed the spec to test the behaviour of the code - that it adds
a `publish_tasks` method to the manual - replacing the previous spec
which tested that the `load` method on
ManualPublishTaskAssociationMarshaller was called. This makes the spec
a bit less brittle in face of future refactoring.

The `it returns the built manual` spec is now double-tested so it has
also been removed.