Remove definition & call to local runTests() from Jenkinsfile
A function of the same name exists in the `govuk` namespace and it is
called from the corresponding stage. The only difference is that our
custom function sets the `TEST_COVERAGE` env var, but this was only
introduced recently and there are probably better ways to do it. I've
captured this difference in a TODO comment and I plan to address it in
a later commit.

I'm hoping to use `govuk.buildProject()` to standardize the Jenkins CI
build and to remove a bunch of unnecessary custom code in the
`Jenkinsfile`. I'm planning to make changes to the `Jenkinsfile` step
by step to document why particular bits of code are no longer necessary.

Note: These changes will not leave the `Jenkinsfile` in a working state
until I've made most/all of them. This commit is just one of those
steps.

References:

* https://github.com/alphagov/govuk-puppet/blob/c439fe32a2c0aebf2b0941f3cab4960d2f183ae5/modules/govuk_jenkins/files/var/lib/jenkins/groovy_scripts/govuk_jenkinslib.groovy#L167-L169