Remove call to govuk.contentSchemaDependency() from Jenkinsfile
As well as the related call to `govuk.setEnvar()`.

These are called by `govuk.buildProject()` in a separate stage.

I'm hoping to use `govuk.buildProject()` to standardize the Jenkins CI
build and to remove a bunch of unnecessary custom code in the
`Jenkinsfile`. I'm planning to make changes to the `Jenkinsfile` step
by step to document why particular bits of code are no longer necessary.

Note: These changes will not leave the `Jenkinsfile` in a working state
until I've made most/all of them. This commit is just one of those
steps.

References:

* https://github.com/alphagov/govuk-puppet/blob/c439fe32a2c0aebf2b0941f3cab4960d2f183ae5/modules/govuk_jenkins/files/var/lib/jenkins/groovy_scripts/govuk_jenkinslib.groovy#L117-L120