Remove call to checkout() from Jenkinsfile
This seems to have been replaced by a call to
`govuk.checkoutFromGitHubWithSSH()` in the corresponding stage within
`govuk.buildProject()`.

I'm hoping to use `govuk.buildProject()` to standardize the Jenkins CI
build and to remove a bunch of unnecessary custom code in the
`Jenkinsfile`. I'm planning to make changes to the `Jenkinsfile` step
by step to document why particular bits of code are no longer necessary.

Note: These changes will not leave the `Jenkinsfile` in a working state
until I've made most/all of them. This commit is just one of those
steps.

References:

* https://github.com/alphagov/govuk-puppet/blob/c439fe32a2c0aebf2b0941f3cab4960d2f183ae5/modules/govuk_jenkins/files/var/lib/jenkins/groovy_scripts/govuk_jenkinslib.groovy#L99-L101
* https://github.com/alphagov/govuk-puppet/pull/5518
* https://github.com/alphagov/govuk-puppet/pull/5551
* https://github.com/alphagov/govuk-puppet/pull/5616