Remove call to govuk.initializeParameters() from Jenkinsfile
Given that `IS_SCHEMA_TEST` and `SCHEMA_BRANCH` are both defined within
`govuk.buildProject()` and the code which uses them is also inside that
function, I strongly suspect we don't need the call to this function.

The above is also possibly related to the introduction of
`govuk.initializeParameters()`.

I'm hoping to use `govuk.buildProject()` to standardize the Jenkins CI
build and to remove a bunch of unnecessary custom code in the
`Jenkinsfile`. I'm planning to make changes to the `Jenkinsfile` step
by step to document why particular bits of code are no longer necessary.

Note: These changes will not leave the `Jenkinsfile` in a working state
until I've made most/all of them. This commit is just one of those
steps.

References:

* https://github.com/alphagov/govuk-puppet/blob/c439fe32a2c0aebf2b0941f3cab4960d2f183ae5/modules/govuk_jenkins/files/var/lib/jenkins/groovy_scripts/govuk_jenkinslib.groovy#L204-L214
* https://github.com/alphagov/govuk-puppet/pull/5304