Call govuk.buildProject() in Jenkinsfile
A bunch of the previous commits have led me to this point where I can
be relatively confident that `govuk.buildProject()` is doing *all* the
things the original `Jenkinsfile` was doing.

However, I strongly suspect it is doing *more* things and some of those
might be undesirable for this project. I'm going to try running the
CI build for this branch to find out.