Replace stub of SectionAssociationMarshaller#load with spec
This commit tests the behaviour of the `load` method on
SectionAssociationMarshaller within the `ManualRepository` which will
allow us to move the code from the Marshaller into this method and
refactor.

The specs are based on those in
`spec/repositories/marshallers/section_association_marshaller_spec.rb`
and test that `sections=` and `removed_sections=` are called on the
manual.