Remove pointless RepublishManualService::ManualNotFoundError
Translating a `ManualRepository::NotFoundError` into a
`RepublishManualService::ManualNotFoundError` exception is pointless
given that the latter is not specifically rescued in
`RepublishManualService`, `ManualsRepublisher`, or
`bin/republish_manuals`. What's more the translation didn't add any
extra information to the underlying exception.

I've removed the `it "raises an exception"` spec, because there's no
longer any relevant logic to test; the behaviour is now just standard
Ruby behaviour.

I've changed the exception class to be an anonymous sub-class of
`StandardError` in the `it "tells none of the listeners to do anything"`
spec, because there's no longer a specific `rescue` block in the code
under test.

Making this change removes one of the last places that code in
`ManualRepository` is used outside the `Manual` class.