Move ManualRepository::NotFoundError -> Manual
This means we can avoid referring to `ManualRepository` in the
`ApplicationController` `rescue_from` block.

I contemplated adding a test for this, but it strikes me as somewhat
strange behaviour - I would expect a 404 if the manual is not found -
so I decided that it wasn't worthwhile. I have, however, manually
tested that if I hit the URL for an unknown manual I am redirected to
the manuals index page with the relevant flash message.

Making this change removes one of the last places that code in
`ManualRepository` is used outside the `Manual` class.