Inline ManualRepository#store into Manual#save
I've introduced the `collection` & `manual` local variables in order to
keep as much of the code from `ManualRepository#store` intact as
possible in this commit to make it easier to see what's going on.
I intend to simplify this in a later commit.

Similarly I've moved the relevant `describe` block from the
`ManualRepository` spec into the `Manual` spec making as few changes
as possible to the code. I've introduced a new context block to contain
the global setup from the `ManualRepository` spec in an effort to make
this as clear as possible and to facilitate moving the other `describe`
blocks in later commits.

I'm confident I'll be able to remove a bunch of duplication from the
`Manual` spec in later commits.