Improve Manual.find spec
The existing spec for this method was moved from the specs for the
ManualRepository class. Rather than test the functionality using a lot
of mocks, this spec creates a database record and ensures Manual.find
retrieves and returns it. Hopefully this spec will be less brittle as
we continue to refactor.