Add migration to fix Highway Code change note data
This is the first part of the fix for the problem described in issue
 #861.

There are unexpected updates listed for The Highway
Code[1] against 1 March 2017.

We expect to see "Annex 5 Penalties" with a single change described as
"Updated the penalty table to increase penalty points for using a
hand-held mobile phone when driving from 3 to 6.".

We're also seeing:

* Road users requiring extra care (204 to 225)
  * Updated Rule 209 to add an image of the school bus road sign.
* Signals to other road users
  * New section added.
* Signals by authorised persons
  * New section added.
* Road markings
  * New section added.
* Vehicle markings
  * New section added.
* Annex 8. Safety code for new drivers
  * Updated the Pass Plus email address to passplus@dvsa.gov.uk

For all the updates on 1 March 2017 we're seeing two identical change
notes, one with a timestamp of 2017-03-01T09:49:18Z and the other with a
timestamp of 2017-03-01T11:03:18Z.

The problem is due to duplicate "major update" `SectionEdition`s and
their change notes. The presence of these SectionEditions result in
"duplicate" `PublicationLog`s which are sent to the Publishing API as
change notes.

This migration updates the affected `SectionEdition`s to make them minor
changes and removes their changes notes.

We still need to rebuild the `PublicationLog`s and update the content in
the Publishing API but they will happen separately.

[1]: https://www.gov.uk/guidance/the-highway-code/updates