Add tests for Manual.all with load_associations argument
Before this commit the effect of `load_associations` argument
introduced in 5a60cd6a0 was untested. Now that ManualRepository has
been merged into Manual, and we have simplified the specs for the
`all` method it is easier to add a spec to test this behaviour.

I've chosen to add a `with_sections` and `with_removed_sections` trait
to the factories. Because there is no association between editions and
sections the code has to set the `section_ids` and
`removed_section_ids` arrays in an `after` block - we can't use any of
the factory_girl magic to do it.