Remove redundant require statements for services
These classes are all in an immediate sub-directory of the `app`
directory and so are auto-loaded by default.

The code was quite inconsistent as to when we did/did not require
these classes and so making this change makes everything consistent.

I used the following regular expressions to find occurrences:

* `require [\"\']\w+_manuals?_service[\"\']`
* `require [\"\']\w+_sections?_service[\"\']`
* `require [\"\']\w+_attachments?_service[\"\']`