Inline most of VersionedManualRepository#get_manual -> Manual
Note that the only place actually calling this method directly is the
`VersionedManualRepository` "unit" spec, so the choice of
`User.gds_editor` is unimportant.

Note also that I've had to make a couple of private methods public so
they're accessible from `Manual#current_versions`.

We an use the `Manual::NotFoundError` exception class instead of
`VersionedManualRepository::NotFoundError`, because the former is
what is being rescued in `ApplicationController` anyway.