Remove outdated info about removing draft manual section
The script mentioned in this section (`remove_draft_manual_section`) was
removed in 654ca53b35be6c3b547df1be8f3e2e5518e31ec6 after the
functionality had been added to the user interface.

I considered updating the README to explain that this could now be
achieved using the UI but felt that it was somewhat unnecessary.

This addresses issue #946.