Add SectionSlugSynchroniser
Manual sections often get into a state where the slug no
longer reflects the title of the section because the slug
is not updated when the title changes.
This is particularly obvious when editors prefix titles with
a number eg '1. Introduction' then wish to reorder the manual
This class reports on potential amendments to slugs to sync them
with their title, it also calls out to the SectionReslugger to
update unconflicting slugs.
A conflict is classed as an attempt to amend a slug to an existing
section slug. These should still be resolved manually.