Push change note validation down into Section
This way we can use the built-in `ActiveModel` validation functionality
rather than duplicating it as was the case in `ChangeNoteValidator`.

The logic in `ChangeNoteValidator` is now redundant, but I've chosen to
keep the "empty" class around for now to reduce the size of the commit.
I have, however, removed the `ChangeNoteValidator` spec, because I've
written equivalent tests in `Section` spec.

Previously in `SectionsController#preview` the section errors were a
simple `Hash` generated by `ActiveModel::Errors#to_hash` via
`ChangeNoteValidator#errors`. `ActiveModel::Errors#to_hash` does not
include entries in the `Hash` for attribute keys with no errors.
Now that the section errors in the controller action are an instance of
`ActiveModel::Errors`, `section.errors[:body]` returns an empty array
rather than `nil`. Hence the change in the condition.