Inline SectionRepository#store into Section#save
Note that `SectionRepository#store` explicitly returned `self` but that
I'm fairly certain the code wasn't making use of it. As such I'm not
returning `self` from `Section.save`.

I've chosen to rewrite the tests rather than try to port them from
section_repository_spec as I didn't think the existing ones were
particularly good. The previous tests appear to imply some behaviour
that doesn't actually exist in the implementation. The new tests
describe the behaviour as it stands. The difference between the
behaviour being described in the old tests and the actual behaviour
might be indicative of a problem with the code but I'm happy to leave
that for now.

I will remove the now-empty SectionRepository and associated specs in
future commit.