Use Manual.build_manual_for in VersionedManualRepository (part 2)
I've introduced an optional `edition` argument to
`Manual.build_manual_for` which means I can override the edition it uses
to build the manual and thus replace the last call to
`VersionedManualRepository#build_manual_for` with a call to
`Manual.build_manual_for`.

The call to `#get_latest_version_of_sections` is the default behaviour
provided by `Manual.build_manual_for`, but the call to
`#get_published_version_of_sections` does something different. For now
I'm just overriding the default behaviour by setting `manual.sections`
after the call to `Manual.build_manual_for`. It's not ideal, but it is
a step in the right direction and allows me to delete a bunch of
conceptual duplication from `VersionedManualRepository`.