Mark copied sections as cloned...
When a manual is updated in order to amend the public updated at field,
all the manual sections are copied with no amendments, this makes it
difficult for the PublicationLogger to discern which section change notes
are copies of the previous edition and so we may end up with duplicate
notes appearing as inaccuracies in the manual change history.
A recent issue with the Highway Code meant these change notes appeared
under the wrong date and implied the guidance was added at a different
time to legislative changes.
This commit persists a boolean  field to signify a cloned
SectionEdition, the PublicationLogger will ignore these editions and
no PublicationLog record will be created.