Inline SectionEdition.two_latest_versions scope
This effectively reverts this commit [1]. Although that commit was
sensible, I've noticed there is some very similar code in
`VersionedManualRepository#get_published_version_of_sections` and I'm
hoping that by inlining the scope into `Section.find` I'll soon be able
to remove the duplication.

Note that this is not a full revert of the aforementioned commit,
because at that point the code in question was in
`SectionRepository#fetch` vs `Section.find`.

Note also that I've left out the call to `SectionEdition.all`, because
it was unnecessary and made testing even more awkward than it already
was.

[1]: ac3c22e461cf2bd1105d9c44463871f9fcee9b41