Filter edition array in Ruby vs Mongoid query in Section.find
Use `Enumerable#take` vs `Mongoid::Contextual::Mongo#limit` in
`Section.find`. This is to make the code as similar as possible to
that in `VersionedManualRepository#get_published_version_of_sections`.
I'm hoping that in a later commit I'll be able to remove the duplication
between that method and `Section.find`, but first I need to make the
code as similar as possible so I can identify the duplication.