Allow ignoring of draft sections in Manual.build_manual_for
This commit serves two purposes:

1. Reduces duplication
2. Collapses behaviour from `VersionedManualRepository` into `Manual`
   and `Section`

This commit inserts the `drop_while` filter from
`VersionedManualRepository#get_published_version_of_sections` into the
`SectionEdition` query in `Section.find`. By default the filter does not
drop any editions, but if the new optional `published` argument is
`true`, then editions are dropped until we reach an edition in the
"published" state.

Unfortunately I've had to bubble the new optional argument up through
`Manual.add_sections_to_manual` and `Manual.build_manual_for`, but I
don't think it's too awful.

Note also that by using the behaviour in `Manual.add_sections_to_manual`
to build the sections, the latter will have a "proper" `SlugGenerator`,
not the so-called "read-only" one which was supplied in
`VersionedManualRepository#get_published_version_of_sections`. Given
that we haven't seen that exception in production and it's not checked
in the tests, I'm pretty confident the slug generator is never used, so
the change in behaviour shouldn't matter.