Use Section::PreviewService#render instead of SectionRenderer#call
The functionality of SectionRenderer#call is also wrapped in
Section::PreviewService#render. An advantage of this commit is that it
allows SectionRenderer to be removed entirely and its functionality
inlined in the PreviewService. However this change does make me think
that PreviewService is doing too much.

Since the SectionPublishingAPIExporter only seems to require some of
the attributes of the rendered section, there may be some more
simplification we can do later.