Avoid odd use of subject in VersionedManualRepository spec
I found the subject blocks in this spec confusing, because they were
being used to store the result from the method under test. I think it's
clearer to explicitly make the call to the method under test within the
test block rather than in a `subject` or `let` block.

Another motivation for making this change was that I'm about to move
these specs into the `Manual` spec where the default subject would be
different. I think this change will make it easier to understand what's
going on.