Add test for unrecoverable GdsApi errors in PublishManualWorker
I want to log these errors to the Rails log as well as Airbrake. I'm
adding these unit tests first to make it easier to test my new behaviour
in a subsequent commit.

The `abort_task` behaviour is also being (indirectly?) tested in the "A
manual fails to publish from the queue due to an unrecoverable error"
scenario of the publishing-a-manual Cucumber feature but that doesn't
feel like the right place for me to test the logging I want to add.