Introduce SearchIndexAdapter
This is intended to provide the only interface that the application
should use to interact with the Rummager API. The intention is that
the method signatures, return values and exceptions should all use
the application's own domain model and *not* the domain model or
concepts specific to the Rummager API. The choice of the word "Adapter"
in the class name is taken from the "Ports and Adapters Pattern" [1].

I've chosen to use "SearchIndex" and not "Rummager" in the class name,
because the application shouldn't care what search-indexing service we
use.

Another motivation for this change is to replace multiple classes each
of which has a single method (`#call`) with a single class which has
multiple methods. My hope is that this is easier to understand and that
we may be able to remove some duplication in the process.

I plan to inline `RummagerManualWithSectionsExporter#call` and
`RummagerManualWithSectionsWithdrawer#call` in a later commit.

[1]: http://alistair.cockburn.us/Hexagonal+architecture