Inline Section.build to Section.new
Note that I've removed the `manual` double being created in the `.find`
describe block of the section_spec. The examples in that block now use the
`manual` double introduced in the commit titled "Instantiate Sections with
manuals instead of slug generators". This avoids a problem where a
`SlugGenerator` was being instantiated with an unexpected prefix ('manual-slug'
instead of '/guidance/manual-slug').