Rename current_section* -> old_section* in SectionReslugger
I found the original naming confusing, because the meaning of the
"current" section changes as the `#call` method executes.

There was already a method named `#remove_old_section_from_rummager`
which I assume was used, because naming it `#remove_current_section_*`
would have been confusing for the reason mentioned above.

I think the new naming is more consistent, e.g. with the code in
`SectionSlugSynchroniser` and generally easier to understand as "old"
is the opposite of "new".