Rename section -> section_edition in SectionReslugger
I found it confusing that a number of variables and method names implied
they were dealing with an instance of a `Section` when they were in
fact dealing with a instance of `SectionEdition`.

I've only changed the "lower-level" occurrences, because I think the
"higher-level" ones reflect the fact that `Sections` themselves are not
persisted in the database; so in these cases the `SectionEdition` is
standing in as a proxy for the `Section` itself.