Extract SearchIndexAdapter#remove_section
This reduces duplication, but I've made it a public method, because I
want to use it from within the `SectionReslugger`.

I think ideally this wouldn't need to be made public, but in order to
make the `SearchIndexAdapter` the only thing that "talks" to `Rummager`,
I think this is a pragmatic step.

Also ideally I don't think this method would need to take `manual`
argument, but it's currently not easy to obtain a section's manual, so
this will do for now.