Use SearchIndexAdapter#remove_section vs Rummager#delete_document
This makes `SearchIndexAdapter` the *only* thing that "talks" to
`Rummager`.

This change effectively removes some non-obvious duplication in terms of
how the document type and ID required by `Rummager` are generated for
the section.

Note that we have to load the old section *before* the slug is updated
so that it doesn't have the new section edition with the new slug.

I have good confidence that I haven't broken anything in making this
change, because of the extra test I recently added to `SectionReslugger`
spec.

In general, I'm sure the `SectionReslugger` could be better written so
that it only uses higher-level APIs like `SearchIndexAdapter`, but this
is hopefully one small step along that road.